refactor(podcast.details.episode): refactor component
Linked to #33
