+++ public/common/filter/html2plainText.js
@@ -1,3 +1,5 @@
-angular.module('ps.common.filter.html2plainText', [])
+import angular from 'angular';
+
+export default angular.module('ps.common.filter.html2plainText', [])
     .filter('htmlToPlaintext', () => (text) => String(text || "").replace(/<[^>]+>/gm, ''))
     .filter('cleanHtml', () => (text) => String(text || "").replace(' =""', ''));
