+++ public/podcasts/details/episodes/episodes.js
@@ -1,97 +1,14 @@
-class podcastItemsListDirective{
-    constructor() {
-        this.restrict= 'E';
-        this.templateUrl = 'podcasts/details/episodes/episodes.html';
-        this.scope = { podcast : '=' };
-        this.controller = 'podcastItemsListCtrl';
-        this.controllerAs = 'pic';
-        this.bindToController = true;
-    }
-}
-
-class podcastItemsListCtrl {
-
-    constructor($scope, DonwloadManager, PodcastItemPerPage, itemService, playlistService ) {
-        /* DI */
-        this.$scope = $scope;
-        this.DownloadManager = DonwloadManager;
-        this.itemService = itemService;
-        this.playlistService = playlistService;
-
-        this.currentPage = 1;
-        this.itemPerPage = PodcastItemPerPage;
-        this.loadPage();
-
-        this.$scope.$on("podcastItems:refresh", () => {
-            this.currentPage = 1;
-            this.loadPage();
-        });
-
-        this.DownloadManager
-            .ws
-            .subscribe( "/topic/podcast/".concat(this.podcast.id),
-                        (message) => this.onMessageFromWS(message),
-                        $scope);
-    }
-
-    onMessageFromWS(message) {
-        var item = JSON.parse(message.body);
-        var elemToUpdate = _.find(this.podcast.items, { 'id': item.id });
-        _.assign(elemToUpdate, item);
-    }
-
-    loadPage() {
-        this.currentPage = (this.currentPage < 1) ? 1 : (this.currentPage > Math.ceil(this.totalItems / this.itemPerPage)) ? Math.ceil(this.totalItems / this.itemPerPage) : this.currentPage;
-        return this.itemService
-            .getItemForPodcastWithPagination(this.podcast, {size: this.itemPerPage, page : this.currentPage - 1, direction : 'DESC', properties : 'pubdate'})
-            .then((itemsResponse) => {
-                this.podcast.items = this.itemService.restangularizePodcastItem(this.podcast, itemsResponse.content);
-                this.podcast.totalItems = itemsResponse.totalElements;
-            });
-    }
-
-    remove(item) {
-        item.remove()
-            .then(() => this.podcast.items = _.reject(this.podcast.items, (elem) => {return (elem.id === item.id); }))
-            .then(() => this.playlistService.remove(item))
-            .then(() => this.loadPage());
-    }
+import angular from 'angular';
+import HtmlFilters from 'common/filter/html2plainText';
+import PlaylistService from 'common/service/playlistService';
+import PodcastItemsListComponent from './episodes.component.js';
 
 
-    reset(item) {
-        return item.reset()
-            .then((itemReseted) => {
-                var itemInList = _.find(this.podcast.items, { 'id': itemReseted.id });
-                _.assign(itemInList, itemReseted);
-                return itemInList;
-            })
-            .then((itemToRemove) => this.playlistService.remove(itemToRemove));
-    }
 
-    addOrRemoveInPlaylist(item) {
-        this.playlistService.addOrRemove(item);
-    }
-
-    isInPlaylist(item) {
-        return this.playlistService.contains(item);
-    }
-
-    swipePage(val) {
-        this.currentPage += val;
-        this.loadPage();
-    }
-
-    stopDownload(item) {
-        this.DownloadManager.ws.stop(item);
-    }
-    toggleDownload(item) {
-        this.DownloadManager.ws.toggle(item);
-    }
-}
-angular.module('ps.podcasts.details.episodes', [
-    'ps.common.service.playlist',
-    'ps.common.filter.html2plainText'
+export default angular.module('ps.podcasts.details.episodes', [
+    PlaylistService.name,
+    HtmlFilters.name
 ])
-    .directive('podcastItemsList', () => new podcastItemsListDirective())
+    .directive('podcastItemsList', PodcastItemsListComponent.component)
     .constant('PodcastItemPerPage', 10)
-    .controller('podcastItemsListCtrl', podcastItemsListCtrl);
+    .controller('PodcastItemsListCtrl', PodcastItemsListComponent);
